requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
youtube-transcript-api>=0.6.0
python-dotenv>=0.19.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
httpx>=0.24.0
pytest-cov>=4.1.0
google-generative-ai>=0.3.0
requests>=2.31.0
